refactor: move imgix core js wrapper
