+++ packages/gatsby-source-url/src/common/imgix-core-js-wrapper.ts
@@ -8,3 +8,6 @@ export const createImgixClient = (
     () => new ImgixClient(options),
     (e) => (e instanceof Error ? e : new Error('unknown error')),
   );
+
+export type IBuildImgixUrl = ImgixClient['buildURL'];
+export type IBuildImgixSrcSet = ImgixClient['buildSrcSet'];
